% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{AddCompStats}
\alias{AddCompStats}
\title{Add Component Statistics to SDA object}
\usage{
AddCompStats(
  SDAres,
  sdThrsh = 0.04,
  maxscoreThrsh = 20,
  maxloadThrsh = 1,
  redoCalc = T
)
}
\arguments{
\item{SDAres}{A data object of SDA results.}

\item{sdThrsh}{A numeric value representing the standard deviation threshold for filtering the data.}

\item{maxscoreThrsh}{An integer value representing the maximum score threshold for filtering the data.}

\item{maxloadThrsh}{A numeric value representing the maximum load threshold for filtering the data.}

\item{redoCalc}{A logical value indicating whether to recalculate the component statistics.}
}
\value{
A data object of SDA results with added component statistics.
}
\description{
This function adds component statistics to SDA results.
}
