% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{CalculateFilters}
\alias{CalculateFilters}
\title{Add Component Statistics to SDA object}
\usage{
CalculateFilters(SDAres, threshold = 100, plot = T, quantThr = 0.95)
}
\arguments{
\item{SDAres}{SDA results obj}

\item{threshold}{A threshold}
}
\value{
kurtosis analysis
}
\description{
This function adds component statistics to SDA results.
}
