% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{convert mouse to human genes}
\usage{
convertMouseGeneList(x)
}
\arguments{
\item{x}{A vector of mouse genes}
}
\value{
A vector of human genes
}
\description{
convert mouse to human genes
}
