% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{download_SDA}
\alias{download_SDA}
\title{Download SDA Results}
\usage{
download_SDA(outf = NULL, outFileID = NULL)
}
\arguments{
\item{outf}{A character string representing the file path to save the SDA results.}

\item{outFileID}{A numeric often 6 digits representing the file identifier for the SDA results.}
}
\value{
the path where the file is saved is returned.
}
\description{
This function downloads SDA results from a remote location and saves them to a file.
}
