% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{DownloadMetadataForSdaResults}
\alias{DownloadMetadataForSdaResults}
\title{DownloadMetadataForSdaResults}
\usage{
DownloadMetadataForSdaResults(sdaResultOutputId)
}
\arguments{
\item{sdaResultOutputId}{A numeric value often 6 digits long}
}
\value{
A dataframe of metadata
}
\description{
DownloadMetadataForSdaResults
}
