% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{GO_enrichment}
\alias{GO_enrichment}
\title{GO_enrichment}
\usage{
GO_enrichment(
  results = NULL,
  component,
  geneNumber = 100,
  threshold = 0.01,
  side = "N",
  OrgDb = NULL
)
}
\value{
GO results
}
\description{
GO_enrichment
}
