% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchShinySDA}
\alias{launchShinySDA}
\title{Launch ShinySDA.}
\usage{
launchShinySDA(...)
}
\value{
Shiny application.
}
\description{
Launch ShinySDA.
}
\details{
Launch ShinySDA.
}
\keyword{SDA}
\keyword{shiny}
