% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{Plot_CorSDA_Loadings}
\alias{Plot_CorSDA_Loadings}
\title{Plot_CorSDA_Loadings}
\usage{
Plot_CorSDA_Loadings(SDAres = NULL)
}
\arguments{
\item{SDAres}{SDA results}
}
\value{
A pheatmap cor heatmap
}
\description{
This function plots Correlatio Heatmap of the gene loadings
}
