% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{Plot_CS_BoxplotMeta}
\alias{Plot_CS_BoxplotMeta}
\title{Plot_CS_BoxplotMeta}
\usage{
Plot_CS_BoxplotMeta(SDAScores = NULL, MetaDF, MetaSelect = "Population")
}
\arguments{
\item{SDAScores}{SDA score matrix}

\item{MetaDF}{Metadata Dataframe}

\item{MetaSelect}{A facctored feature of Metadata}
}
\value{
A pheatmap cor heatmap
}
\description{
This function plots Correlatio Heatmap of the gene loadings
}
