% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{plot_SDA_ChiSqrHM}
\alias{plot_SDA_ChiSqrHM}
\title{Plot SDA Results Using Chi-Square Heatmap}
\usage{
plot_SDA_ChiSqrHM(MetaDF, MetaSelect, SDAScores, direction, clustStat = T)
}
\arguments{
\item{MetaDF}{A data frame containing metadata for the samples.}

\item{MetaSelect}{A character string representing the metadata column to use for plotting.}

\item{SDAScores}{A data frame containing the SDA scores.}

\item{direction}{A character string indicating the direction of the analysis (either "pos" or "neg").}

\item{clustStat}{A logical value indicating whether to include clustering statistics in the plot.}
}
\value{
A pheatmap plot using a chi-square
}
\description{
This function plots SDA results using a chi-square heatmap.
}
