% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{plotEnrich}
\alias{plotEnrich}
\title{plotEnrich}
\usage{
plotEnrich(
  GeneSetsDF,
  GeneVec,
  plotTitle = "",
  xLab = "",
  N = NULL,
  k = NULL,
  ReturnPval = T,
  BiPlot = F
)
}
\value{
plot of enrichment
}
\description{
plotEnrich
}
