% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{preprocess_SDA}
\alias{preprocess_SDA}
\title{Preprocess SDA Results}
\usage{
preprocess_SDA(
  SDAres = NULL,
  QuantThr = 0.95,
  envv = NULL,
  TopN = 150,
  MetaDF = NULL
)
}
\arguments{
\item{SDAres}{A data frame containing the SDA results.}

\item{QuantThr}{A numeric value between 0 and 1 representing the quantile threshold for filtering the data.}

\item{envv}{An environment or list containing the data used in the SDA analysis.}

\item{TopN}{An integer value representing the number of top features to retain.}

\item{MetaDF}{A data frame containing metadata for the samples.}
}
\value{
A list containing the filtered and summarized SDA results.
}
\description{
This function preprocesses SDA results by filtering and summarizing the data.
}
