% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{print_gene_list}
\alias{print_gene_list}
\title{print_gene_list}
\usage{
print_gene_list(
  results = results,
  i,
  PrintRes = F,
  PosOnly = F,
  NegOnly = F,
  AbsLoad = T,
  TopN = 150
)
}
\value{
ordered gene list
}
\description{
print_gene_list
}
