% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{print_loadings_scores}
\alias{print_loadings_scores}
\title{print_loadings_scores}
\usage{
print_loadings_scores(
  SDAResult = NA,
  ComponentN = NA,
  ColFac = NA,
  Prefix = "SDAV",
  GeneLoc = NA
)
}
\value{
plots of gene loadings
}
\description{
print_loadings_scores
}
