% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{Run_DowLoadSDA_evv}
\alias{Run_DowLoadSDA_evv}
\title{This function to update the environment list when doanloading SDA obj from prime-seq}
\usage{
Run_DowLoadSDA_evv(envv, input)
}
\arguments{
\item{envv}{environment list associated with ShinySDA}

\item{input}{environment list associated with innput paramters in ShinySDA}
}
\value{
the updated environment list
}
\description{
This function to update the environment list when doanloading SDA obj from prime-seq
}
