% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{Run_GeneAnn_evv}
\alias{Run_GeneAnn_evv}
\title{This function to update the environment list when getting gene annotations}
\usage{
Run_GeneAnn_evv(envv, input)
}
\arguments{
\item{envv}{environment list associated with ShinySDA}

\item{input}{environment list associated with innput paramters in ShinySDA}
}
\value{
the updated environment list
}
\description{
This function to update the environment list when getting gene annotations
}
