% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{Run_MetaDF_local_evv}
\alias{Run_MetaDF_local_evv}
\title{This function to update the environment list when grabbing the meta data DF}
\usage{
Run_MetaDF_local_evv(envv, input, session)
}
\arguments{
\item{envv}{environment list associated with ShinySDA}

\item{input}{environment list associated with innput paramters in ShinySDA}

\item{session}{session  associated with ShinySDA}
}
\value{
the updated environment list
}
\description{
This function to update the environment list when grabbing the meta data DF
}
