% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.R
\name{SDA_ChiSqrHMresDF}
\alias{SDA_ChiSqrHMresDF}
\title{Plot SDA Results Using Chi-Square Heatmap}
\usage{
SDA_ChiSqrHMresDF(MetaDF, MetaSelect, SDAScores)
}
\arguments{
\item{MetaDF}{A data frame containing metadata for the samples.}

\item{MetaSelect}{A character string representing the metadata column to use for plotting.}

\item{SDAScores}{A data frame containing the SDA scores.}
}
\value{
A list of two vector text of enrichments, one for pos one for neg scored cells
}
\description{
This function plots SDA results using a chi-square heatmap.
}
